Refactorización y optimización de lógica de datos y UI
Se han realizado los siguientes cambios principales:

- Refactorización de `ObtenerArticulos` y `ObtenerClientes` para devolver `DataTable` y añadir métodos específicos para cargar datos en `DataGridView`.
- Cambio de tipo de `decimal` a `float` en variables numéricas (`totalVenta`, `montoTotal`, etc.) en `Form1.cs` y `FormFacturacion.cs`.
- Actualización de `CargarProductos` en `Form1.cs` para usar `Articulo.ObtenerArticulos` en lugar de consultas SQL directas.
- Simplificación de la lógica de configuración de encabezados en `UtilidadesUI.cs`, eliminando código redundante.
- Manejo explícito de valores nulos en `detalles.cs` al cambiar el tipo de retorno de `UtilidadesBD.BuscarRegistro`.

Estos cambios mejoran la modularidad, reutilización y claridad del código, además de optimizar el manejo de datos y la presentación en la interfaz de usuario.

+++ SistemaVentas/FormFacturacion.cs
@@ -9,11 +9,11 @@ namespace SistemaVentas
     public partial class FormFacturacion : Form
     {
         private DataGridView dgvProductos;
-        private decimal montoTotal;
-        private decimal subtotal;
-        private decimal descuentoAplicado;
+        private float montoTotal;
+        private float subtotal;
+        private float descuentoAplicado;
 
-        public FormFacturacion(DataGridView productos, decimal total)
+        public FormFacturacion(DataGridView productos, float total)
         {
             InitializeComponent();
             
@@ -102,7 +102,7 @@ namespace SistemaVentas
                     txtDescuento.Text = "0";
                 }
 
-                descuentoAplicado = subtotal * (decimal)(descuentoPorcentaje / 100);
+                descuentoAplicado = subtotal * (float)(descuentoPorcentaje / 100);
                 montoTotal = subtotal - descuentoAplicado;
 
                 lblDescuentoValor.Text = $"-${descuentoAplicado:F2}";
